# server/openapi.yaml
openapi: 3.0.0
info:
  title: Communicator API 2026
  description: >
    API REST + WebSocket para sistema de mensajería instantánea y
    llamadas peer-to-peer (WebRTC). Soporta chat 1-a-1, grupos,
    presencia en línea, y señalización de llamadas de voz/video.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor Local de Desarrollo (ADB reverse)
  - url: http://192.168.XX.XX:3000
    description: Red WiFi local (reemplazar XX.XX con IP del servidor)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en /api/auth/login o /api/auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Credenciales inválidas
...
    Message:
      type: object
      required: [id, sender_id, receiver_id, group_id, content, message_type, created_at, read_at, reply_to_id]
      properties:
        id:
          type: integer
          example: 42
        sender_id:
          type: integer
          example: 1
        receiver_id:
          type: integer
          nullable: true
          example: 2
        group_id:
          type: integer
          nullable: true
          example: null
        content:
          type: string
          example: Hola, ¿cómo estás?
        message_type:
          type: string
          enum: [text, image, system]
          default: text
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
          nullable: true
        sender_name:
          type: string
          example: juan_perez
        sender_avatar:
          type: string
          nullable: true
        group_name:
          type: string
          nullable: true
          description: Nombre del grupo (solo para mensajes grupales)
        reply_to_id:
          type: integer
          nullable: true
          example: 12
          description: ID del mensaje al que responde
        replied_to:
          type: object
          nullable: true
          required: [id, sender_id, sender_name, content, message_type]
          properties:
            id:
              type: integer
              example: 12
            sender_id:
              type: integer
              example: 1
            sender_name:
              type: string
              example: juan_perez
            content:
              type: string
              example: Mensaje original
            message_type:
              type: string
              enum: [text, image, system]
          description: Mensaje original al que se responde (incluido cuando reply_to_id no es null)
...
    MessagesResponse:
      type: object
      required: [messages, hasMore]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        hasMore:
          type: boolean
          description: true si hay más páginas disponibles
...
    Group:
      type: object
      required: [id, name, description, avatar_url, created_by, created_at, role]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          maxLength: 100
          example: Amigos del trabajo
        description:
          type: string
          nullable: true
          maxLength: 255
          example: Grupo para coordinar proyectos
        avatar_url:
          type: string
          nullable: true
        created_by:
          type: integer
          example: 1
        workspace_id:
          type: integer
          nullable: true
          description: Workspace al que pertenece el grupo (los grupos se aislan por workspace activo)
          example: 2
        created_at:
          type: string
          format: date-time
        role:
          type: string
          enum: [admin, member]
          description: Rol del usuario autenticado en el grupo
    GroupMember:
      type: object
      required: [id, username, avatar_url, is_online, role]
      properties:
        id:
          type: integer
        username:
          type: string
        avatar_url:
          type: string
          nullable: true
        is_online:
          type: boolean
        role:
          type: string
          enum: [admin, member]
    CreateGroupRequest:
      type: object
      required: [name, memberIds]
      properties:
        name:
          type: string
          maxLength: 100
          example: Amigos del trabajo
        description:
          type: string
          maxLength: 255
          example: Grupo para coordinar proyectos
        memberIds:
          type: array
          minItems: 1
          items:
            type: integer
          description: IDs de los miembros a agregar
    CreateGroupResponse:
      type: object
      required: [id, name, description]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
...
  responses:
    Unauthorized:
      description: Token JWT requerido o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: No eres miembro de este grupo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/groups:
    post:
      summary: Crear un grupo
      description: |
        Crea un nuevo grupo. El creador se asigna automáticamente como admin.
      tags: [Grupos]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Grupo creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupResponse'
        '400':
          description: Nombre y miembros requeridos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      summary: Listar grupos del usuario
      description: Grupos a los que pertenece el usuario autenticado, ordenados alfabéticamente.
      tags: [Grupos]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de grupos
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/groups/{id}:
    get:
      summary: Obtener un grupo por ID
      description: Devuelve la información de un grupo específico.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Información del grupo
          content:
            application/json:
              schema:
                type: object
                properties:
                  group:
                    $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Grupo no encontrado

    put:
      summary: Actualizar nombre/descripción del grupo
      description: Actualiza el nombre y la descripción del grupo. Solo el administrador puede realizar esta acción.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Nuevo nombre del grupo
                description:
                  type: string
                  nullable: true
                  description: Nueva descripción del grupo
      responses:
        '200':
          description: Grupo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupResponse'
        '400':
          description: Nombre del grupo obligatorio
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/groups/{id}/members:
    get:
      summary: Listar miembros de un grupo
      description: Devuelve todos los miembros del grupo con su rol y estado online.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del grupo
      responses:
        '200':
          description: Miembros del grupo
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupMember'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Agregar miembros al grupo
      description: Solo el administrador puede agregar miembros.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del grupo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                memberIds:
                  type: array
                  items:
                    type: integer
                  minItems: 1
              required:
                - memberIds
      responses:
        '200':
          description: Miembros agregados exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: integer
                    description: Número de miembros agregados
        '400':
          description: Se requiere al menos un miembro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/groups/{id}/members/{userId}:
    delete:
      summary: Eliminar un miembro del grupo
      description: Solo el administrador puede eliminar miembros. No puede eliminarse a sí mismo.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del grupo
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID del usuario a eliminar
      responses:
        '200':
          description: Miembro eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Miembro eliminado
        '400':
          description: No puedes eliminarte a ti mismo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: El usuario no es miembro del grupo
  /api/groups/{id}/members/me:
    delete:
      summary: Salir del grupo
      description: Cualquier miembro puede salirse del grupo.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del grupo
      responses:
        '200':
          description: Saliste del grupo exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Has salido del grupo
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Grupo no encontrado
  /api/groups/{id}/transfer:
    post:
      summary: Transferir propiedad del grupo
      description: |
        El creador del grupo transfiere la propiedad a otro miembro.
        El nuevo propietario obtiene rol admin, el anterior pasa a miembro.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del grupo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: integer
                  description: ID del nuevo propietario
      responses:
        '200':
          description: Propiedad transferida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Propiedad transferida exitosamente
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Grupo no encontrado
  /api/groups/{id}/messages:
    get:
      summary: Obtener mensajes de un grupo
      description: |
        Mensajes del grupo con paginación por offset.
        Orden descendente por fecha, entregados en orden ascendente.
      tags: [Grupos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID del grupo
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: Mensajes del grupo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ══════════════════════════════════════════════
  # MÓDULO DE LLAMADAS (signaling + pending)
  # ══════════════════════════════════════════════
...